+++ climpred/tests/test_bootstrap.py
@@ -447,46 +447,3 @@ def test_bootstrap_p_climatology(hindcast_hist_obs_1d, metric):
             .all()
         )
     assert bskill.sel(skill=reference, results="p").sel(lead=lead)[v] < 0.1
-
-
-@pytest.mark.parametrize(
-    "reference", [None, "climatology", "uninitialized", "persistence"]
-)
-@pytest.mark.parametrize("alignment", ["same_verif", "same_inits", "maximize"])
-def test_bootstrap_valid_time(hindcast_hist_obs_1d, alignment, reference):
-    result = hindcast_hist_obs_1d.bootstrap(
-        iterations=ITERATIONS,
-        metric="rmse",
-        comparison="e2o",
-        dim=[],
-        alignment=alignment,
-        reference=reference,
-    )
-    assert "time" not in result.dims
-    assert "valid_time" in result.coords
-    assert len(result.coords["valid_time"].dims) == 2
-    if reference:
-        assert set(result.dims) == set(["init", "lead", "results", "skill"])
-    else:
-        assert set(result.dims) == set(["init", "lead", "results"])
-
-
-@pytest.mark.parametrize(
-    "reference", [None, "climatology", "uninitialized", "persistence"]
-)
-@pytest.mark.parametrize("alignment", ["same_verif", "same_inits", "maximize"])
-def test_verify_valid_time(hindcast_hist_obs_1d, alignment, reference):
-    result = hindcast_hist_obs_1d.verify(
-        metric="rmse",
-        comparison="e2o",
-        dim=[],
-        alignment=alignment,
-        reference=reference,
-    )
-    assert "time" not in result.dims
-    assert "valid_time" in result.coords
-    assert len(result.coords["valid_time"].dims) == 2
-    if reference:
-        assert set(result.dims) == set(["init", "lead", "skill"])
-    else:
-        assert set(result.dims) == set(["init", "lead"])
